Fix sequence number formatting (need i64 to hold u32 and -1)

+++ algorithm/phantom_grow/AwayMessage.h
@@ -13,7 +13,7 @@ typedef nx_struct AwayMessage
 
 } AwayMessage;
 
-inline int32_t Away_get_sequence_number(const AwayMessage* msg) { return msg->sequence_number; }
+inline int64_t Away_get_sequence_number(const AwayMessage* msg) { return msg->sequence_number; }
 inline int32_t Away_get_source_id(const AwayMessage* msg) { return msg->source_id; }
 
 #endif // SLP_MESSAGES_AWAYMESSAGE_H
